added basic arithmetic operations example
