+++ hello.js
@@ -1,82 +1,14 @@
-//Console Input
-const readline=require("readline");
+console.log("Hello World");
 
-const rl = readline.createInterface({
-    input:process.stdin,
-    output: process.stdout
-})
-function mainMenu(){
-console.log("---------------------------");
-console.log("Factorial Application");
-console.log("1. Say Hello");
-console.log("2. Factorial");
-console.log("3. Exit Application");
-console.log("---------------------------");
+num1 = 2;
+num2 = 3;
 
-rl.question("Enter your choice (1-3): ", choice =>{
-    console.log(choice);
+sum = num1 + num2;
+diff = num1-num2;
+prod = num1 * num2;
+quo = num1 / num2;
 
-    if (choice === "1"){
-         sayHello();
-    } else if (choice === "2"){
-          computeFactorial();
-    } else if (choice === "3"){
-        exitProgram();
-    } else {
-        console.log("Invalid Choice. Please Try Again");
-    }
-
-    })
-    }
-     
-    function sayHello(){
-    console.log("Hello");
-    }
-
-    
-   function computeFactorial(){
-    console.log("Factorial");
-
-        rl.question("Please enter a number for factorial: ", numStr =>{
-            let num = parseInt(numStr);
-
-            if (isNaN(num) || num < 0){
-                console.log("Please enter a non-negative integer");
-            } else {
-                let fact = 1;
-
-                for (let i=1; i <= num; i++){
-                    fact *=i;
-                }
-
-                console.log("The factorial of " + num + " is " + fact);
-            }
-              backToMenu();
-        })
-}
-function exitProgram(){
-    console.log("Exiting Application");
-    rl.close();
-    console.clear();
-}
-function backToMenu(){
-    console.log("--------------");
-    console.log("1. Back to menu");
-    console.log("2. Exit");
-{
-     rl.question("What would you like to do next? (1-2) ", (backToMenuChoice)=>{
-
-        if(backToMenuChoice==="1"){
-            console.clear();
-            mainMenu();
-        } else if (backToMenuChoice ==="2"){
-            exitProgram();
-        } else {
-            console.log("Invalid Choice. Please Try Again!");
-            backToMenu();
-        }
-
-    })
-}
-}
-mainMenu();
+console.log("The sum of "+num1+" & "+num2+" is "+sum);
+console.log("The difference of "+num1+" & "+num2+" is "+diff);
+console.log("The product of "+num1+" & "+num2+" is "+prod);
+console.log("The quotient of "+num1+" & "+num2+" is "+quo);
